requirements.txt


Flask
Flask-Login
python-dotenv
pandas
scikit-learn
joblib
requests
mysql-connector-python
